# 3 <==> 4 Channel Shuffling with Hardware Intrinsics