--- a/src/ImageSharp/Common/Helpers/Shuffle/IComponentShuffle.cs
+++ b/src/ImageSharp/Common/Helpers/Shuffle/IComponentShuffle.cs
@@ -30,12 +30,19 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct DefaultShuffle4 : IComponentShuffle
     {
-        public DefaultShuffle4(byte p3, byte p2, byte p1, byte p0)
-            : this(SimdUtils.Shuffle.MmShuffle(p3, p2, p1, p0))
-        {
-        }
+        private readonly byte p3;
+        private readonly byte p2;
+        private readonly byte p1;
+        private readonly byte p0;
 
-        public DefaultShuffle4(byte control) => this.Control = control;
+        public DefaultShuffle4(byte p3, byte p2, byte p1, byte p0)
+        {
+            this.p3 = p3;
+            this.p2 = p2;
+            this.p1 = p1;
+            this.p0 = p0;
+            this.Control = SimdUtils.Shuffle.MmShuffle(p3, p2, p1, p0);
+        }
 
         public byte Control { get; }
 
@@ -44,12 +51,11 @@ namespace SixLabors.ImageSharp
         {
             ref byte sBase = ref MemoryMarshal.GetReference(source);
             ref byte dBase = ref MemoryMarshal.GetReference(dest);
-            SimdUtils.Shuffle.InverseMmShuffle(
-                this.Control,
-                out int p3,
-                out int p2,
-                out int p1,
-                out int p0);
+
+            int p3 = this.p3;
+            int p2 = this.p2;
+            int p1 = this.p1;
+            int p0 = this.p0;
 
             for (int i = 0; i < source.Length; i += 4)
             {
@@ -63,7 +69,9 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct WXYZShuffle4 : IComponentShuffle
     {
-        public byte Control => SimdUtils.Shuffle.MmShuffle(2, 1, 0, 3);
+        private static readonly byte WXYZ = SimdUtils.Shuffle.MmShuffle(2, 1, 0, 3);
+
+        public byte Control => WXYZ;
 
         [MethodImpl(InliningOptions.ShortMethod)]
         public void RunFallbackShuffle(ReadOnlySpan<byte> source, Span<byte> dest)
@@ -89,7 +97,9 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct WZYXShuffle4 : IComponentShuffle
     {
-        public byte Control => SimdUtils.Shuffle.MmShuffle(0, 1, 2, 3);
+        private static readonly byte WZYX = SimdUtils.Shuffle.MmShuffle(0, 1, 2, 3);
+
+        public byte Control => WZYX;
 
         [MethodImpl(InliningOptions.ShortMethod)]
         public void RunFallbackShuffle(ReadOnlySpan<byte> source, Span<byte> dest)
@@ -112,7 +122,9 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct YZWXShuffle4 : IComponentShuffle
     {
-        public byte Control => SimdUtils.Shuffle.MmShuffle(0, 3, 2, 1);
+        private static readonly byte YZWX = SimdUtils.Shuffle.MmShuffle(0, 3, 2, 1);
+
+        public byte Control => YZWX;
 
         [MethodImpl(InliningOptions.ShortMethod)]
         public void RunFallbackShuffle(ReadOnlySpan<byte> source, Span<byte> dest)
@@ -135,7 +147,9 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct ZYXWShuffle4 : IComponentShuffle
     {
-        public byte Control => SimdUtils.Shuffle.MmShuffle(3, 0, 1, 2);
+        private static readonly byte ZYXW = SimdUtils.Shuffle.MmShuffle(3, 0, 1, 2);
+
+        public byte Control => ZYXW;
 
         [MethodImpl(InliningOptions.ShortMethod)]
         public void RunFallbackShuffle(ReadOnlySpan<byte> source, Span<byte> dest)