--- a/src/ImageSharp/Common/Helpers/Shuffle/IPad3Shuffle4.cs
+++ b/src/ImageSharp/Common/Helpers/Shuffle/IPad3Shuffle4.cs
@@ -65,9 +65,11 @@ namespace SixLabors.ImageSharp
 
     internal readonly struct XYZWPad3Shuffle4 : IPad3Shuffle4
     {
-        private static readonly byte XYZW = SimdUtils.Shuffle.MmShuffle(3, 2, 1, 0);
-
-        public byte Control => XYZW;
+        public byte Control
+        {
+            [MethodImpl(InliningOptions.ShortMethod)]
+            get => SimdUtils.Shuffle.MmShuffle(3, 2, 1, 0);
+        }
 
         [MethodImpl(InliningOptions.ShortMethod)]
         public void RunFallbackShuffle(ReadOnlySpan<byte> source, Span<byte> dest)