… src/ImageSharp/Common/Helpers/SimdUtils.HwIntrinsics.cs
… src/ImageSharp/Common/Helpers/SimdUtils.Shuffle.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Argb32.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Bgr24.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Bgra32.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Rgb24.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Rgba32.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/_Common.ttinclude
… src/ImageSharp/PixelFormats/Utils/PixelConverter.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/FromVector4.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/FromVector4_Rgb24.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/Pad3Shuffle4Channel.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/Shuffle3Channel.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/Shuffle4Slice3Channel.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/ShuffleFloat4Channel.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/ToVector4_Rgb24.cs
… tests/ImageSharp.Tests/Common/SimdUtilsTests.Shuffle.cs
… tests/ImageSharp.Tests/Common/SimdUtilsTests.cs