@antonfirsov 2020-11-02
@JimBobSquarePants I want to do a proper review here, but not sure about the timing. Lucky case: by Wednesday evening, worst case: by Sunday. Is this acceptable for you?