@JimBobSquarePants 2020-11-02
@antonfirsov No worries. Hopefully someone else from the team can also have a look in the interim. 

I added the Rgb24/Vector4 benchmarks to https://github.com/SixLabors/ImageSharp/issues/1354#issuecomment-720636039

We're looking at a very healthy speedup on all target frameworks.