@antonfirsov 2020-11-02
@JimBobSquarePants I tried to destile some derived numbers to make conclusions from https://github.com/SixLabors/ImageSharp/issues/1354#issuecomment-720636039: 

### ToVector4_Rgb24

Count    | Runtime | Master | Branch | Speedup (`Master/Branch`)
-- | -- | -- | -- | --
64 | .NET 4.7.2 | 310.2 | 355.5 | 0.87257384
64 | .NET Core 2.1 | 230.2 | 228.5 | 1.007439825
64 | .NET Core 3.1 | 236.7 | 217 | 1.09078341
256 | .NET 4.7.2 | 622.5 | 448.9 | 1.386723101
256 | .NET Core 2.1 | 498.1 | 309.2 | 1.610931436
256 | .NET Core 3.1 | 436.8 | 212.3 | 2.05746585
2048 | .NET 4.7.2 | 3,460.60 | 1,974.10 | 1.753001368
2048 | .NET Core 2.1 | 3,421.30 | 1,985.50 | 1.723142785
2048 | .NET Core 3.1 | 2,972.20 | 1,165.00 | 2.551244635

### FromVector4_Rgb24

Count  | Runtime | Master | Branch | Speedup (`Master/Branch`)
-- | -- | -- | -- | --
64 | .NET 4.7.2 | 316.4 | 320.8 | 0.986284289
64 | .NET Core 2.1 | 238.9 | 246 | 0.971138211
64 | .NET Core 3.1 | 250.3 | 243.4 | 1.028348398
256 | .NET 4.7.2 | 1,051.30 | 967 | 1.087176836
256 | .NET Core 2.1 | 846.8 | 1,003.30 | 0.844014751
256 | .NET Core 3.1 | 640.2 | 437 | 1.464988558
2048 | .NET 4.7.2 | 4,551.20 | 4,391.60 | 1.036342108
2048 | .NET Core 2.1 | 4,390.40 | 4,225.60 | 1.039000379
2048 | .NET Core 3.1 | 2,979.40 | 1,822.70 | 1.634607999

#### Things I don't understand:
- Why do the benchmarks show speedup for 2.1 and .NET Framework for `Count >= 256`? There is no SIMD, and it doesn't look like we changed something there.
- Is there a regression for `FromVector4_Rgb24` for 2.1 with `Count == 256`? (noise?)