@JimBobSquarePants 2020-11-02
@antonfirsov the speedup is due to this change. Our original pet pixel implementation turned out to be quite slow. 

https://github.com/SixLabors/ImageSharp/pull/1409/files#diff-92b35c7d3a5d6901602e57d9dfd618b0d6ceb8cf3442fcfa044111c0e1f509dfR73

The regression is just noise. There’s massive error on that run for some reason