@peter-dolkens 2020-11-03
The "Fastfallback" is a regression in performance - presumably because it is no longer falling back, but doing it properly, quickly.

Is this only ever triggered algorithmically? 

Could people be running this in fallback mode deliberately as it's faster and close enough for their needs?

Should you be retaining the "Fastfallback" path, or happy to discontinue that support.

For simplicity, I'd certainly lean towards taking the hit, but you know your library better than anyone.

Not sure if it's the kind of thing that might deliberately get used in low-quality bulk tasks like thumbnail generation.