@JimBobSquarePants 2020-11-04
@peter-dolkens if you check the linked benchmarks the error for that regression run is super high. Likely due to my laptop throttling. I’ll post updated ones in the morning